Implement abstractions for VAO's and VBO's
VertexBuffer keeps track of a VBO and it's data.

BufferLayout refers to a VertexBuffer and describes it's layout; it does
so with a BufferElement array. It also keeps track of the stride.

BufferElement corresponds to an array of vertex atribute data. This
could be a position or texturecoördinates for example. It keeps track of
how many components the vertex attribute data consists of (count), the
name of the attribute variable whose location is to be queried in the
shader by OpenGL (name), and the offset from the first component of the
first vertex attribute in the layout (offset).

VertexArray keeps track of a VAO and a BufferLayout array. When you use
it with a shader it manages all required VBO's and maps the vertex
attributes to the shader veriables using the BufferLayout array and
stores it in the VAO. Binding the VertexArray then binds the VAO. When
you use the VertexArray with a shader the VertexArray will also be
bound.

Other changes include:

 * Remove the depend target and rule in both Makefiles
 * Change the code formatting in engine/window.h
 * Add a TODO list to game.c
 * Remove logging of shader compilation
 * Remove some commented-out code