Fix delta_time calculation
My suspicions of delta_time being calculated incorrectly were correct. I
was using process time instead of the real time. This means that I was
using the time the process used the cpu, which is less than the actual
time. It still laggs when you get close to the cube though.