New vector library and small stylistic changes
The old vector library was an overengineered piece of trash that
deserves to burn in hell. I'm not sure if I should be ashamed that I
wrote that shit, or proud that I was able to make such a complex thing.

Anyway, I have finally started working in this engine again and I hope I
will continue to do so. Next up on my list is loading obj files and
lighting.