Implement loading OBJ files and rendering them
For now it can only load OBJ files without colours or textures. In the
future I'll create new variations of the mesh and vertex struct for
different types of meshes such as those that do have colour and/or
textures. I will also create functions for easily drawing those meshes
and default shaders to go along with them. I will also need to implement
some sort of lighting (probably a directional light to begin with).