Modify OBJ loading and mesh code to handle normals
Normals are required for lighting in the shader. I changed the mesh
structure and added a face structure. I am storing the normals per-face
instead of per-vertex. This way I am sort of converting per-vertex to
per-face and back to per-vertex. The reason is that I think you should
be able to easily create a mesh manually.